% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_point.R
\name{coords_to_point}
\alias{coords_to_point}
\title{Create a Spatial Vector from coordinates}
\usage{
coords_to_point(x, y)
}
\arguments{
\item{x}{x coordinate}

\item{y}{x coordinate}
}
\value{
A point Spatial Vector
}
\description{
Create a Spatial Vector from coordinates
}
\author{
Ignacio Ramos-Gutiérrez
}
