% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampletree}
\alias{exampletree}
\title{Example phylogeny}
\format{
A phylogenetic tree with 120 tips and 119 internal nodes
\describe{
Phylogenetic tree with 120 tips.
}
}
\usage{
exampletree
}
\description{
Example phylogeny for phyloshapeR package
}
\keyword{datasets}
