% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_territory.R
\name{get_territory}
\alias{get_territory}
\title{Select a country's focal territory based on size}
\usage{
get_territory(shape, order = 1)
}
\arguments{
\item{shape}{Spatial Vector including only one country}

\item{order}{which territory to select, based on size. order=1 selects the biggest one, order=2 the second, etc.}
}
\value{
A plot
}
\description{
Select a country's focal territory based on size
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
world <- terra::vect(phyloshapeR::world)
australia <- world[world$name=="Australia",] 
mainland.aus <- get_territory(australia, 1)
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutiérrez
}
