% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getshape.R
\name{getshape}
\alias{getshape}
\title{Create a predefined shape}
\usage{
getshape(type = "polygon", sides = 6, rotate = 0)
}
\arguments{
\item{type}{Type of predefined shape. Accepted values are "polygon", "heart", "diamond", "star", triangle", "square", "hexagon", "circle"}

\item{sides}{Applies if type is "polygon". Number of sides of the polygon}

\item{rotate}{Degrees in which to rotate the shape (counter-clockwise)}
}
\value{
A SpatVector object.
}
\description{
Create a polygonal or custom prediefined shape
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 shape <- getshape("polygon", sides=6, rotate = 360/12)
 terra::plot(shape)
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutiérrez
}
