% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomap.R
\name{phylomap}
\alias{phylomap}
\title{Default plot for phyloshape plotting}
\usage{
phylomap(tree, offset = 0, cex = 0.85, pch = 16, tip.label.col = "black", ...)
}
\arguments{
\item{tree}{Phylogenetic tree obtained from phyloshape function}

\item{offset}{separation from the tree tip to the tip label point}

\item{cex}{Point size of the tree label point}

\item{pch}{Point character of the tree label point}

\item{tip.label.col}{Colour of the tree label point}

\item{...}{Other arguments tu pass through plot.phylo() function for the phylogeny plotting}
}
\value{
A plot
}
\description{
Plot the phyloshape in the default mode, with some editable options
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 shape <- getshape("polygon", sides=6, rotate = 360/12)
 tree <- phyloshape(shape,  nlines = 100)
 phylomap(tree, cex=0.4, offset=10000)  
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutiérrez
}
