% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloshape.R
\name{phyloshape}
\alias{phyloshape}
\title{Create tree to plot}
\usage{
phyloshape(
  shape,
  tree = NULL,
  point = NULL,
  ntips = 360,
  method = "fill",
  depth.k = 0.95
)
}
\arguments{
\item{shape}{a SpatialVector object with only one geometry whose contour will be used to shape the phylogeny}

\item{tree}{Phylogenetic tree to midify its branch lengths. If NULL, a tree of random topology will be created.
The number of tips will be randomly pruned to match 'ntips'}

\item{point}{a SpatialVector point from where the phylogeny will be plotted. If NULL, the function will use the centroid.}

\item{ntips}{number of lines (tips) the resulting phylogeny will include. default is 360.}

\item{method}{Method used to calculate the branch lengths. Method 'extend' will mantain
the original phylogeny shape and add distances just to the tip branches.
Method 'fill' will calculate distances from the tips back to the root. Distances from
a tip or node to its direct ancestor are calculated using a depth constant.}

\item{depth.k}{Applies for method 'fill'. Percentage of branch that separates internal nodes.
Values near to 0 have clustered internal nodes near to the root, thus resemble to a polytomy,
while values near 1 have internal nodes clustered toward tips. Default value is 0.95.}
}
\value{
A phylogenetic tree whose edge lengths will draw the specified shape.
}
\description{
Create phylogenetic tree with edge lengths according to a shape
}
\details{
Argument \code{depth.k} stands for the amount of shared branch two nodes share. Therefore, a \code{depth.k} = 0
means they join at the root (which would return a polytomy), while a value of 1 would mean the branch
is common until the depth of the deepest node of them both (i.e. splitting at the tip).
Values lower than 0.9 will generally result in phylogenies with nodes clustered near to the root.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 tree2 <- phyloshape(tree=NULL, shape,  method = "fill", ntips = 150, depth.k = 0.99)
 ape::plot.phylo(tree2, type = "f", show.tip.label = F)
 ape::tiplabels(pch=16, cex=.5)
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutiérrez
}
