% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{world}
\format{
A terra PackedSpatVector object
\describe{
A terra PackedSpatVector object with world borders.
}
}
\usage{
world
}
\description{
Example data to create attendance certificates
}
\keyword{datasets}
