% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{world}
\format{
An object of class \code{PackedSpatVector} of length 1.
}
\usage{
world
}
\description{
Example data to create attendance certificates
}
\details{
A terra PackedSpatVector object.

\if{html}{\out{<div class="sourceCode">}}\preformatted{}\if{html}{\out{</div>}}
}
\keyword{datasets}
